chore: move to parser directory
